ENGA-1243: created getRandomId, throttle in utils
