+++ src/typeFiles/js/utils/throttle.js
@@ -1,0 +1,24 @@
+export default function throttle(func, waitTime) {
+    let isThrottled = false,
+        savedArgs,
+        savedThis;
+
+    function wrapper() {
+        if (isThrottled) {
+            savedArgs = arguments;
+            savedThis = this;
+            return;
+        }
+        func.apply(this, arguments);
+        isThrottled = true;
+        setTimeout(function() {
+            isThrottled = false;
+            if (savedArgs) {
+                wrapper.apply(savedThis, savedArgs);
+                savedArgs = savedThis = null;
+            }
+        }, waitTime);
+    }
+
+    return wrapper;
+}
